move fake block gas ut to app_test.go
